deleting of course from learning journey

+++ Backend/Routes/Dependent/CourseLearningJourneyRoutes.py
@@ -14,7 +14,7 @@ from random import randint
 @app.post("/courselearningjourney/seedall", tags=["CourseLearningJourney", "SeedAll"])
 def seedAll():
     try:
-    # get learning journey
+        # get learning journey
         session = Session(engine)
         getLearningJourneyStatement = select(LearningJourneyModel)
         allLearningJourneys = session.exec(getLearningJourneyStatement).all()
@@ -43,6 +43,7 @@ def seedAll():
             "success": False,
             "messaage": str(e)
         }
+
 
 @app.delete("/courselearningjourney/deleteall", tags=["CourseLearningJourney", "DeleteAll"])
 def deleteAll():
@@ -96,31 +97,39 @@ async def AddCourseLearningJourney(request: Request, session: Session = Depends(
             "message": errors
         }
 
+
 @app.delete("/courselearningjourney/", tags=["CourseLearningJourney"])
 async def deleteCourseLearningJourney(request: Request, session: Session = Depends(get_session)):
     errors = []
     try:
         # find the course
         requestData = await request.json()
-        selectedCourseStatement = select(CourseModel).where(CourseModel.Course_ID == requestData["Course_ID"])
+        selectedCourseStatement = select(CourseModel).where(
+            CourseModel.Course_ID == requestData["Course_ID"])
         selectedCourse = session.exec(selectedCourseStatement).all()
-        if(len(selectedCourse) == 0):
-            errors.append("Couse does not exist!")
-        elif(len(selectedCourse) == 1):
+        if (len(selectedCourse) == 0):
+            errors.append("Course does not exist!")
+        elif (len(selectedCourse) == 1):
             errors.append("Only one course left cannot delete!")
         else:
             selectedCourse = selectedCourse[0]
 
-
         # delete all the relations
         # hard delete the relations
         allRelatedRelationStatement = select(CourseLearningJourneyModel).where(
-            CourseLearningJourneyModel.Course_ID == requestData["Course_ID"])
+            CourseLearningJourneyModel.Course_ID == requestData["Course_ID"] and CourseLearningJourneyModel.LearningJourney_ID == requestData["LearningJourney_ID"])
         relationResults = session.exec(allRelatedRelationStatement)
         allRelations = relationResults.all()
+        if len(allRelations) == 0:
+            errors.append("Course not inside learning journey!")
+        if len(errors) > 0:
+            return {
+                "success": False,
+                "message": errors
+            }
         for relation in allRelations:
             session.delete(relation)
-            
+
         session.commit()
         session.close()
         return {
